feat: 특정 위치 기준 근처 대여소 검색 기능 구현
- 근처 대여소 검색을 위한 엔드포인트 `/api/v1/stations/nearby` 추가
- 특정 위치(위도, 경도)값을 기준으로 가까운 대여소 반환
- `radius`, `limit`으로 검색 반경, 검색 개수 제한 가능
- DB 내장 함수 및 Point 타입 사용을 위한 테이블 구조 변경

+++ db/init.sql
@@ -8,7 +8,7 @@ DROP TABLE IF EXISTS station;
 DROP TABLE IF EXISTS location_group;
 
 CREATE TABLE location_group (
-                                group_id    INT             NOT NULL    AUTO_INCREMENT,
+                                group_id    BIGINT             NOT NULL    AUTO_INCREMENT,
                                 group_name  VARCHAR(50)     NOT NULL,
                                 PRIMARY KEY (group_id)
 ) ENGINE=InnoDB
@@ -16,7 +16,7 @@ CREATE TABLE location_group (
 CREATE TABLE station (
                         stn_id      VARCHAR(50)     NOT NULL,
                         stn_no      VARCHAR(50)     NULL,
-                        group_id    INT             NULL,
+                        group_id    BIGINT             NULL,
                         stn_name    VARCHAR(100)    NOT NULL,
                         stn_addr1   VARCHAR(255)    NULL,
                         stn_addr2   VARCHAR(255)    NULL,
